Add the missing `tabspaces-exclude-buffers`
But there is still a problem.

I'm using consult with tabspaces, so I used the config inside
README. But when I want to add `*Messages*` to
`tabspaces-include-buffers`. It doesn't work when I'm using
`switch-to-buffer`, because `*Messages*` is not inside
`(frame-parameter nil 'buffer-list)` and `(frame-parameter nil
'buried-buffer-list)` when Emacs initialized.

After I opened `*Messages*` and bury it, it goes to `(frame-parameter
nil 'buried-buffer-list)`. While `tabspaces--local-buffer-p` only
checks `(frame-parameter nil 'buffer-list)`, `*Messages*` will not
show inside `consult--source-workspace`.

Does the only solution is to use the README's

```emacs-lisp
(when (get-buffer "*Messages*")
  (set-frame-parameter nil
                       'buffer-list
                       (cons (get-buffer "*Messages*")
                             (frame-parameter nil 'buffer-list))))
```

I'm not saying this is not good, but it looks a bit ugly to me...